+++ perl/FSM/Adapter/ISF/Parser.pm
@@ -148,6 +148,7 @@ sub _build_actor($self, $actor_ast, $source_label) {
     }
 
     $self->_validate_rule_trigger_targets($result);
+    $self->_validate_rule_priority_targets($result);
 
     fsm_trace_exit('Parser _build_actor completed', 3);
     return $result;
@@ -373,6 +374,27 @@ sub _validate_rule_trigger_targets($self, $actor) {
                 unless defined($target)
                     && !ref($target)
                     && $transaction_names{$target};
+        }
+    }
+
+    return 1;
+}
+
+sub _validate_rule_priority_targets($self, $actor) {
+    my $actor_name = $actor->{actor_name};
+    my %rule_names = map { $_->{name} => 1 } @{$actor->{rules} || []};
+
+    for my $rule (@{$actor->{rules} || []}) {
+        my $rule_name = $rule->{name};
+        for my $action (@{$rule->{actions} || []}) {
+            next unless ref($action) eq 'ARRAY' && @$action;
+            next unless defined($action->[0]) && !ref($action->[0]) && $action->[0] eq 'priority';
+
+            my $target = $action->[2];
+            confess "Error: rule '$rule_name' priority targets unknown rule '$target' in actor '$actor_name'\n"
+                unless defined($target)
+                    && !ref($target)
+                    && $rule_names{$target};
         }
     }